requirements.txt
Django
python_dotenv
google-generativeai
qdrant_client
djangorestframework
pypdf
pypdf2
python-docx
cohere